Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-2
> 
> PulkoMandy reports there already was a rebuilt HaikuWebKit against ICU 66 which this version of the script apparently failed to pick up. So clearly it needs some work.

To keep track of infos that was only on irc: actually, I had bumped the revision and triggered the rebuild, but it turns out build with icu66 is broken.

I have now created a 1.7.1 tag on the webkit repo (with this buildfix and some other bugfixes that don't need changes to haiku builders), and will push the recipe soon. The script did the right thing.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>