WebPositive: Realigned Settings window
The idea is to eventually allow the user to choose between a bunch of predefined search engines and optionally allowing them to enter their own search engine, should they want to do so. For that, a greyed out text input field, which will be ungreyed whenever the user opts for using their own search engine, will be required. The 'Search page' option will eventually be the most space-consuming option, so it'd make sense from my perspective to put it in the bottom of that list, while making it more distinct and grouping together all options with a text field in the meantime.