Update patch set 1
Patch Set 1:

I found this which is possibly better:

https://stackoverflow.com/a/10841376

With some changes for gcc2 support, I got something like this which seems to work fine:

#define ioctl(a, b, c, d...) ioctl(a, b, c, (0, ## d ))

No need for the extra function and I think this is slightly less complicated to read (even if still not great).

Note that gcc2 preprocessor seems a bit picky about the whitespace in the macro definition.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>