Update patch set 1
Patch Set 1:

> Nice, does this work for "ioctl(fd, B_IOCTL_GRANT_TTY);" ?
> It fails here to build libroot.

No, it can add only one argument. I didn't know we had cases with only the command and no arguments at all.

For reference, on FreeBSD, what they do is encode the number of arguments into the command (using the _IO macro to define the commands). This allows their ioctl implementation to know how many arguments there are. But, it appears that in BeOS, the same ioctl number can mean different things for different devices (see https://github.com/jscipione/r5headers/search?q=ioctl where several drivers define B_DEVICE_OP_CODES_END+1 and following opcodes) so we can't even do that.

Here is another option that also works with two arguments, but requires a separate function:

struct ioctl_args{
    void* c;
    size_t d;
};

void _ioctl_with_args(int a, int b, struct ioctl_args x)
{
    _kern_ioctl(a, b, x.c, x.d)
}

#define ioctl(a, b, c...) _ioctl_with_args(a, b, (struct ioctl_args){ c })

int main(void)
{
    int foo;
    ioctl(1, 2);
    ioctl(1, 2, &foo);
    ioctl(1, 2, &foo, 4);
}

The unspecified fields in the structure are defaulted to 0.

There is one limitation: this requires the type of c and d to always be a pointer and a size_t, or something which can be casted to that without warnings. Are there ioctls which work differently? (it still works, but there is a warning and I don't think we can silence it).

Patch-set: 1