Update patch set 2
Patch Set 2: Code-Review+2

We would need a quite complicated macro in C to manage to have ioctl() called with 2, 3, or 4 arguments routed to the right functions. This solutions is much simpler to understand.

It's unfortunate we can't use C++ here, it would indeed make things easier (just have default args to the function and no need for varargs). But we have to provide this API in C, and once we have done that, I see no point in having a separate implementation for C++.

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+2