Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Moved into new src/add-ons/network path, add-on search paths updated.
> 
> instantiate_uri_addon is in GopherRequest.cpp (this behavior matches the decor add-ons and several other "non-kernel" server/kit plugins)
> 
> > Performance wise, it seems not great to rescan the directory here for every request.
> >
> > I'd rather scan it once initially on instanciation of the protocol roster, and node monitor it. Load all the add-ons then.
> > 
> > And this method should then be something like addonmap[protocol]->MakeRequest(...);
> 
> I agree, but didn't think there was anything "persistent" since this all exists within the network kit and not the net_server?

Having it in the net server would make it persistant for the whole system. I did not ask for that. Just make it persistant for the application lifetime.

- Application starts: it scans the add-ons (in BNetworkRoster constructor, for example) and loads them (in an std::map fronm the protocol name to the instanciation method, for example)
- A request is created: it just looks up the correct add-on for the protocol in the map, and calls its instanciate method

This is the same approach we use for media decoding, for example. add-ons/media is scanned and all add-ons loaded once. Then you can play as many files as you want, reusing the add-on. Nothing is done in the media server for the decoding/encoding part (it used to a few years ago, but not anymore).

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>