Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 1:
> 
> I like LRUCache better, too, or LRUHashMap ;-)
> 
> In any case, I missed how the implementation is used within HaikuDepot. I also wouldn't mind to move the LRUCache to shared/ instead.

There was already an icon cache in the form of a plain HashMap. That one (AFAICT) did not limit the number of icons stored though, which the LRU one fixes. They're API compatible, so the only change needed was changing the type of the class member.

Patch-set: 2
CC: Gerrit User 1000015 <1000015@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>