BSecureSocket: remove one hardcoded path and add a FIXME about another.
Trying to get BSecureSocket and direct users of OpenSSL to converge to
the same default set of certificates. Untested for now, I'm working on
something else and this is a drive-by fix. I'll check if that work when
I'm done with other things and am able to reboot to a test Haiku
install, unless someone wants to check before then.