Update patch set 14
Patch Set 14:

(2 comments)

I'm not sure if the semantics of this class are that nice, and intuitive: shouldn't Write() block (or return an error like EMSGSIZE) until the complete buffer could be written? At least that's how I would expect it to work. The way it works now is that it can really only be used as a byte stream, not a message stream.

Patch-set: 14
Reviewer: Gerrit User 1000016 <1000016@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>