Update patch set 4
Patch Set 4:

Yes, as things are going I think we will be very similar to x86 or x86_64. A lot of weird things was probably due to u-boot or missing compiler flags. Atm I use a x86 linker script with changed output, and maybe .ini_array things and it seems to work for loading..

Patch-set: 4