Update patch set 1
Patch Set 1:

Does this actually work? I remember trying to do this in OpenTracker long ago, and due to the way the poseview was using CopyBits() to move list rows around as they were being sorted, the end result was the background image being turned into a complete mess for a folder with a large number of items.

Patch-set: 1
CC: Gerrit User 1000015 <1000015@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>