Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Problem is likely in BTextView introduced in hrev54549, not Tracker. My MsgDump utility is also broken in similar way after that change. Old BeOS software ma be also broken.

This particular bug is unique to Tracker.

There are some changes in BTextView in hrev54549 that may cause regressions in apps where the text rect dimensions are set differently than the text view dimensions. I believe that the Vision bug #16568 is an example of this. The BTextView changes are intentional to match the behavior on BeOS, however, I'm not sure if that is the cause of your issue in MsgDump or not, it's possible that there is still a bug in BTextView, or it's possible that I could change the BTextView behavior to fix the bug ignoring BeOS app compatibility.

Can you point me to the code for your MsgDump utility so that I can see if there is a discrepancy there? This app could provide a useful test case to help me to locate any bugs in  my BTextView implementation. StyledEdit has been a great help in locating BTextView bugs as it is a wrapper around BTextView and thus exposes bugs related to changes in wrapping and alignment. I've tested quite a few BeOS apps to look for regressions there. Can you point me to a BeOS app that has an issue with text view?

Thank you for your feedback.

Patch-set: 1
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>