Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> While I agree that there should be only one place for this, haven't we already changed the settings file location of OpenSSL already? Wouldn't it then make sense to also "haiku-ify" the certificate path there as well? (Yeah, I know I'm late to the party)

We can, the provided sample config file already was not matching the actual locations (it uses cacerts.pem instead of CARootCertificates.pem, and now it also has the wrong path for the user certificates). It is packaged in /system/data/ssl/opnssl.cnf.

When is it used? Only by the openssl command line tool? It seems libssl/libcrypto don't use it, otherwise no certificates would be found.

Patch-set: 4