Update patch set 1
Patch Set 1:

What's the status with AboutSystem with this?

Part of what got us into these changes to BTextControl is that AboutSystem initially started with a 1px wide text rect. While inserting text to it, this led to adding linebreaks between every character, which was very slow and delayed the application start.

Now I see a 1px text rect making a comeback and I'm wondering if it would reintroduce that problem. Did you test it in that case?

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>