Update patch set 2
Patch Set 2:

I wonder why I would ever want to have MediaPlayer re-order the dropped files? Shouldn't it always use the sort-order used in the Tracker window the files were dropped from?

It does work when selecting a few songs in a folder and hitting enter. They are opened as a playlist in the Tracker window's order.

I'd say if we wanted reordering, a columnlistview for the playlist would be nicer.

Or am I missing something?

BTW, possibly better than a modifier-key-drag would be a context menu on right-drag.

Patch-set: 2
CC: Gerrit User 1000004 <1000004@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>