Update patch set 23
Patch Set 23: Code-Review+1

(2 comments)

Can we merge this? The test is not great (and there are some comments on it) but I think the main code is fine now.

This change is somewhat important for security, it will allow for example gcrypt to make sure no private crypto data is accidentally stored on disk because the memory is swapped out.

Patch-set: 23
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+1