BWindow: Add PopUpMenusBeginning() and PopUpMenusEnded()
These hook methods are called just before a BPopUpMenu is opened
or just after it is closed like MenusBeginning() and MenusEnded()
but instead they work on asynchronous BPopUpMenu's.

New methods were created so that we preserve backwards compatibility
rather than repurposing MenusBeginning() and MenusEnded().

In BPopUpMenu check to see if we are attached to a BMenuBar first.
If we are, most likely because we're in a BMenuField BMenuBar will
call MenusBeginning() and MenusEnded() instead. Use private static
convenience method _HasMenuBar() to check this.

Add new private message constant _POP_UP_MENUS_DONE_ to AppDefs.h
just after _MENUS_DONE_. This message constants is used to send
a message back to BWindow to tell it that the menu has closed from
the menu window thread. There are no equivalent constant for
PopUpMenusBeginning() or MenusBeginning() because those methods are
called from the window thread in BPopUpMenu and BMenuBar.

Add FBC compatibility macros for the new methods.

Change-Id: I0161272f968db419e4b2887ddebbef90810183c4