IK: Add PopUpMenusBeginning() and PopUpMenusEnded()
These new BWindow hook methods are called just before a BPopUpMenu is
opened or just after it is closed as well as being called just before a
menu attached to a BMenuBar is opened or just after it is closed. These
methods serve as replacements for MenusBeginning() and MenusEnded()
that work for both a BMenu attached to a BMenuBar as well as an
asynchronous BPopUpMenu that is not attached to a BMenuBar. New methods
were created so that we preserve backwards compatibility.

In BPopUpMenu check to see if we are attached to a BMenuBar first. If we
are, (e.g. in a BMenuField) BMenuBar will call PopUpMenusBeginning()
and PopUpMenusEnded() for us. Create private static class method
_HasMenuBar() for this purpose.

Update documentation for MenusBeginning() and MenusEnded() marking these
methods as deprecated and add documentation for the new methods.

Add new private message constant _POP_UP_MENUS_DONE_ to AppDefs.h
just after _MENUS_DONE_. This message constants is used to send a message
to BWindow that the menu has closed from the menu window thread. There
are no equivalent message constant for PopUpMenusBeginning() or
MenusBeginning() because those methods are called directly from the
window thread in BPopUpMenu and BMenuBar.

Add FBC compatibility macros for these new methods.

Call PopUpMenusBeginning() and PopUpMenusEnded() in Tracker instead of
MenusBeginning() and MenusEnded() so that these methods are called for both
regular menus and pop-up menus.

Change-Id: I0161272f968db419e4b2887ddebbef90810183c4