Update patch set 18
Patch Set 18:

Unfortunately, this patch set is still not quite yet ready. It live updates the menu items when you press or release the shift key when a submenu is not open, but when you have a submenu open it does not update the menu items. I'm not sure yet how I can make this work with submenus open but I wanted to update the patch set so that it no longer has a merge conflict.

Patch-set: 18
Reviewer: Gerrit User 1000013 <1000013@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>