Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2:
> > > 
> > > > Patch Set 2:
> > > > 
> > > > > Patch Set 2:
> > > 
> > > This works for submenus, but not for supermenus. When a submenu is open, it doesn't relayout the parent menu and live-updating shiftable Tracker menus don't update.
> > 
> > Actually... this patch set does allow a pop-up menu to update with a submenu open, but does not update a menu bar menu to update with submenu open... so it's really close to working.
> 
> It work on pop-up menus with submenu open SOME of the time... so there's something going on that I don't yet fully understand.

Sorry to post so many comments here.

fSelected in BMenu is not set in _UpdateWindowViewSize() when a menu is open, so this patch set doesn't work. fSelected should point to the selected menu item. We need to figure out a better way to find the selected menu item here or set fSelected.

Patch-set: 2