Update patch set 7
Patch Set 7:

> Patch Set 7: Code-Review+2
> 
> Still not too sure about the locking but no one remembers how it was supposed to work apparently, so if we didn't spot a deadlock, let's assume it's doing the right thing...

If I understood it right (remember I'm not a developer), the locks on the FontCacheEntries are twofold. First, to protect their glyph caches as you said on your comment on PS3. Read to use them, write when creating glyphs (thus adding them to the cache). Then FontCacheEntry::CreateGlyph notes that you need a write-lock for both the main font and the fallback[0] when you use one. If it was only for the glyph caches you shouldn't need to lock the fallback. From what I can see, it's used there because the creation of the glyph is done in two phases: PrepareGlyph and WriteGlyphTo, with properties stored on the font engine itself. Maybe those can be extracted to a different object making the engine stateless with respect to that and not needing that lock, but that's something for a pro to analyze and do if worth it.

There's also the FontManager lock when getting the styles for the fallbacks. I didn't find out why I had to release the lock on the FontCacheEntry before, but I tried not to do it and got deadlocks playing like crazy with the sliders in FontDemo. And I'd have to release it some times anyway to assure ordered locking.

[0] https://git.haiku-os.org/haiku/tree/src/servers/app/font/FontCacheEntry.cpp#n288

Patch-set: 7
Reviewer: Gerrit User 1000223 <1000223@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>