Update patch set 1
Patch Set 1:

> > > Thanks for the patch! Could you please merge the cleanup patch and this one before committing them? It doesn't make much sense to split this into two commits.
> > 
> > So sorry for the mess :S Second patch with gerrit, I did not expect that. I'm going to fix indentation and re-read submitt guideline.

It's not too much of a problem, and partially our fault for deciding to use a process quite different from what you see in other places like Github (not to mention using Git in the first place, which isn't really an easy tool).

You can configure git to do "auto-rebase" if you don't want it to create merge commits all the time, I think this is more suitable for working in Haiku:
https://coderwall.com/p/tnoiug/rebase-by-default-when-doing-git-pull

(this suggests putting it in the global conf with the --global option, but you can also omit that to enable it only in the Haiku repo if you prefer to).

> Another tip: make sure when you amend/update your commit(s), that the message contains the change id you find above (Change-Id: Ib0b7de56ef35092da93cc5fa61bdf0cfef41c042). This will make sure that when you submit an updated patch, it will be added to this change and not create a new one.

In fact you shoud probably install the commit hook as documented here:
https://review.haiku-os.org/Documentation/cmd-hook-commit-msg.html

It generates the change-id for you when you create a new commit.

For these already existing ones, you will need to add it manually, however. But it will be useful for your next submissions.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>