Update patch set 11
Patch Set 11: Code-Review+1

(2 comments)

I think we should merge this.

While I agree that the current versiom results in being a bit verbose, I think it is a good idea to have ownership clearly indicated in the code, and as long as we are stuck with gcc2, we can't use std::unique_ptr.

What we could do, however:
- Add an IsNull (or IsValid?) method to avoid writing ".Get() == NULL" in many places.
- Maybe add an operator= for setting an ObjectDeleter from a pointer

It's true that in the current state, there are still raw pointers to objects exposed in various places in the API. But now we can more easily see where this happens (by looking for calls to .Get()) and I think it makes the code a bit easier to analyze. And now we can work from there to see if we can reduce the usage of raw pointers further, and redesign things to make sure there are no more memory management problems.

As the patch is already quite large, it makes sense to first merge this part, and then do further changes later.

Patch-set: 11
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+1