Update patch set 2
Patch Set 2:

> It can be done if it is fine to create an instance of DeleteFunc at each call. This is no-op for current AutoDeleter usage, but in theory DeleteFunc can have private state.

I don't see a practical use for that. The deleter function is created by the AutoDeleter, its constructor has no arguments, so where would the state come from?

And if you need to store some state, you can create an AutoDeleter subclass to put it in there (which is what we usually do).

Patch-set: 2