Update patch set 2
Patch Set 2:

Hi,

It is preferred to include the Change-Id in your commit messages, it's this line listed in Gerrit: Change-Id: I07d4cedbd23f3e9e950bef0409fdb941c2c32a20

This way Gerrit notices that the new commit is an update of the previous one and automatically replaces it, instead of making a new one.

It's preferrable to do this because this way, all review comments and the complete history of the patch are accessible in the same place.

You can set up the Gerrit commit hook to add this change id automatically for you so you don't forget about it.

Patch-set: 2