Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review-1
> 
> The symbol should however be kept in libroot.so for binaries until beta2 (public API compatibility)

Why? It is not documented and not present in BeOS. It should be no 3rd-party software that use it (but it will be good to check out of tree drivers like OSS). It seems to be introduced by mistake.

Patch-set: 3
Reviewer: Gerrit User 1000172 <1000172@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>