Update patch set 3
Patch Set 3: -Code-Review

> Why? It is not documented and not present in BeOS. It should be no 3rd-party software that use it (but it will be good to check out of tree drivers like OSS). It seems to be introduced by mistake.

The API is also available to applications (not only drivers) and was in the public headers of Haiku for 15 years. It's not unlikely that someone used it outside Haiku.

So, removing it from the .h files first (to break the build, but not usage of existing binaries) at least for one beta version makes sense.

Patch-set: 3
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0