Update patch set 3
Patch Set 3:

> Is it possible to declare symbol as alias to unload_driver_settings?

This is how it was done initially. But it created problems when building on OS X (because this code is built for the host OS) had been removed in c7cb9b2f595838104e085124b79cd7983aa11d34.

We can check the target with a #ifdef to know if the file is being built for Haiku or some other OS, and have the alias only available for Haiku.

Patch-set: 3
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=0