Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> I likewise fail to see the point in the filename changes. "This is how everyone else does it" isn't a particularly good justification.

> Patch Set 2: Code-Review 2
>
> I still do not know what you are getting at here. We have some files in the root of the repository with a case and organization slightly different than other projects, but still acceptably so; again, GitHub detects these files just fine. What exactly is "fossilized" about this?

I don't mean to be the contradictor that wastes everyone's here, but just because everyone else does it, that doesn't mean automatically that it's bad. Latching onto what we have now without good reason is what's bad here, particularly because the original decision in the first place that led to what we have to today was most probably made based on the exact same question; "How does everyone else do that?". I described the current format as a relic because it's a product of its time, and I tried to make the point that we also have to evolve accordingly.

I'm also concerned that the current document is what I've heard described as a "cognitohazard" to some people, which could render this into an accessibility issue, but I haven't managed to ask around and confirm that just yet, so please take that with a grain of salt for the time being because I haven't looked into it too much.

Patch-set: 2
Reviewer: Gerrit User 1000173 <1000173@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>