Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Using double instead of float can reduce efficiency.
> 
> Very probably, but I'd go for correctness. Just create a 32 bit PCM file with a clean sine wave using full amplitude and listen to it.
> 
> There are other options, though. One could keep the float calculations and then only convert to double for the clamping, or even just use as maximum value the one that is <= INT32_MAX while still exactly represented as float. Both of those would need to add clamping to the sites that don't have it, and I don't know how much different the first one would be.
> 
> Open for changes, of course.

Floats only have 24 bits of precision so they are not the right type to use here. I think the best thing to do is use int32 in all cases, except when converting to or from float, in that case, use float. It should be doable with template specialization.

Patch-set: 1
CC: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>