Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > Floats only have 24 bits of precision so they are not the right type to use here.
> 
> If I understand correctly, when human hear louder sound, perceived precision will decrease. High dynamic range is more important than absolute precision.

The difference is not something you can hear because soundcards only have 16 or maybe 18 bits of output. But it's not only about that, since the mixer can be instanciated in other places in the audio processing chain. It should do its job in a way as lossless as possible because we don't know how the output will be post-processed.

In fact we don't even know if the output is audio meant for humans to listen. The Media Kit can be used as a generic media processing tool, and any 1D waves could be mixed here.

Patch-set: 1