Update patch set 1
Patch Set 1:

OK, this discussion is way over of what I can understand, so I'll leave it for the experts.

I want to say one thing, though, in case I've not been clear in the description. This is not about loss of sound precision. If I understand the mixer, it always converts input to float, does the mixing and then converts to whatever the output needs, so more precision would need wider changes. This is instead about a wrap around of values when converted back to int. The maximum positive int32 (and a few others) becomes 2^31 as float, and that back to int32 becomes -2^31, the most *negative* value. That is, a complete change, not just a loss of a bunch of the least significant bits.

Patch-set: 1
Reviewer: Gerrit User 1000223 <1000223@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>