Update patch set 1
Patch Set 1:

(1 comment)

It would typically help to add a specific unit test for the input values which exhibit the issue (possibly for each of the function modified in this change). See my comment in Resampler.cpp.
Otherwise yes, the compiler seems to generate a different code only for the integer case.

Patch-set: 1
CC: Gerrit User 1000023 <1000023@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>