Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> About B_LAST_BUFFER_ERROR: did someone check BeOS behavior?

I haven't.

When reading a media track that uses a raw decoder instead of ffmpeg (in Haiku, not BeOS) I get all the frames and then the error (though it's a general one instead of LAST_BUFFER), but I have not installed BeOS to test.

Even if I think the correct behaviour is to be able to get all frames, be they an integer multiple of the buffer size or not, there's no hurry to merge this. We may be taking about much less than the last tenth of a second of a track, which probably is silence or near it anyway.

Patch-set: 2
Reviewer: Gerrit User 1000223 <1000223@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>