Update patch set 2
Patch Set 2:

> When reading a media track that uses a raw decoder instead of ffmpeg (in Haiku, not BeOS) I get all the frames and then the error (though it's a general one instead of LAST_BUFFER), but I have not installed BeOS to test.
> 
> Even if I think the correct behaviour is to be able to get all frames, be they an integer multiple of the buffer size or not, there's no hurry to merge this. We may be taking about much less than the last tenth of a second of a track, which probably is silence or near it anyway.

The question is only if you get both the last buffer and the error at the same time, or if you get the last buffer with B_OK, and then the last buffer error with no data in the next call. In either case you should of course get the last buffer (and the current behavior is wrong).

The buffer size may be quite large, and result in 1/2 or 1/4 of a second being missing, possibly not that important when listening to a full song, but more annoying for a short sound effect.

I can check BeOS behavior when I get back home where I have a BeOS installed on a computer

Patch-set: 2
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>