Update patch set 2
Patch Set 2:

Looking closer at the mixer code, it seems that it would make sense to use AUDIO_INT as the internal format for the mixer instead of AUDIO_FLOAT.

The reason is we assume normalized floats (in the range -1 to 1), so we don't make use of the extra dynamic of floats. Float only has a precision of 24 bits so it cannot represent arbitrary int32 values.

I think the loss of precision from converting float to int32 is more acceptable, as it is loss of data that wouldn't be of any use in actual audio output.

That's a bit more work than the changes proposed here but I think it makes more sense.

Patch-set: 2