Update patch set 3
Patch Set 3: Code-Review-2

I think this makes things worse in that it hides the problem instead of straight out crashing as would happen with a division by 0.

The result will probably be that things will continue to not work, but will now be harder to debug.

If you're really worried about a divide by 0, an assert would be ok (but would still crash at the same place, with a better error message)

As Maximo commented, the code does various multiplications and divisions that you simplified, but it results in code that's harder to understand because the local variable store values in strange units (hundreds of frames per seconds? who uses that?).

So, the code is more complex, the failure isn't really fixed but most likely moved elsewhere further from the driver (which would be the problem source here), and that's for a failure that doesn't even happen, to start with.

Patch-set: 3
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=-2