Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > The reason is we assume normalized floats (in the range -1 to 1), so we don't make use of the extra dynamic of floats.
> > 
> 
> Your point about the 24 (maybe 25 with the "hidden bit"?) still stands, but as the mixer mixes, the buffer is in fact in -nInputs to nInputs range. And you apply float gains here and there, so the wrapping issue would still exist with int32 as mix buffer format.

It's 23 + the not-encoded bit for a total of 24.

The buffer is int*_min to int*_max for integers and converting from one integer type to another is not too hard without a risk of overflow (it can be done without any multiplication, with some bitshifts). But it is easier to do with unsigned values.

For float, we need to do some clamping however, since it can potentially be anywhere outside the range.

Patch-set: 2