Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > [...] but as the mixer mixes, the buffer is in fact in -nInputs to nInputs range. And you apply float gains here and there, so the wrapping issue would still exist with int32 as mix buffer format.

I've failed again to make myself clear, sorry.

There are three steps in the mixer:
1. Resample input to mixer format
2. Add the inputs in the mixer buffer
3. Resample mixer buffer to output format

All of them apply gains.
1 may change sample rate in addition to format.
Notice that the mixing in step 2 may very well produce values out of [-1,1], that later in 3 will be scaled by whatever gain and clamped to the output range (bugs aside).

We currently have a problem in 3 if the output format is int.

Naively[1] using int as the mixer format we would still have the same problem in 3, as we have to scale by float gain anyway. And we would also have it in 1, as it should still work with float input. And we would have another one in 2, as we may be adding several inputs.

[1] Naive is as far as I can go. Very probably someone knows how to do all of that safely, efficiently, in an uncomplicated manner and without rewriting everything, but that's not me.

Patch-set: 2
Reviewer: Gerrit User 1000223 <1000223@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>