Update patch set 2
Patch Set 2:

> There are three steps in the mixer:
> 1. Resample input to mixer format
> 2. Add the inputs in the mixer buffer
> 3. Resample mixer buffer to output format

> 1 may change sample rate in addition to format.

Yes, so the mixer can work in a single format and sample rate, making it a lot simpler.

> Notice that the mixing in step 2 may very well produce values out of [-1,1], that later in 3 will be scaled by whatever gain and clamped to the output range (bugs aside).

The advantage of float here is that it allows to go outside the normal range. The problem is it is only a 24bit format, so there is some loss of precision if you are outputting to int32.

The additional problem is that the rounding error can lead us to have something outside the range that can be represented in the output.

> Naively[1] using int as the mixer format we would still have the same problem in 3, as we have to scale by float gain anyway. And we would also have it in 1, as it should still work with float input. And we would have another one in 2, as we may be adding several inputs.

Yes, I missed the fact that step 2 works with things outside the -1;1 range, so using something other than float is a problem there.

Applying a float gain is not really a problem on its own, it's possible to do an int by float multiplication and convert the result back to int, as long as you make sure the output is in the correct range.

So, it would be something like this:

int32 someValue;
float gain;
float multiplied = someValue * gain;

if (multiplied > INT32_MAX)
    output = INT32_MAX;
else
    output = (int32)multiplied;

I think something like this is better than what you suggested, as it does a better (but still not perfect) job at letting the signal pass unmodified in the setup where you have one int32 input, and one int32 output, and all gains set to 1 (that should be an easy case, right?). In the current situation we can do this only in step 3.

We still have an additional problem: if you have a high end audio output that can handle 32bit int format, and an high auqlity file in 32bit format, the mixer will still convert the data to float for processing, losing 8 bits of precision in the process.

So, yes, we could rewrite everything to:
- Step 1: apply resampling, input gain, format ocnversion, and output gain to each input
- Step 2: add all the results from step1 together, with an overflow check.

By having the output gain done in step 1, we optimize things a bit (one step is saved) and in step 2, we make sure the result is already in the normalized range -1 .. 1, or at least that clamping it to that range is the correct thing to do.

Patch-set: 2