Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Applying a float gain is not really a problem on its own, it's possible to do an int by float multiplication and convert the result back to int, as long as you make sure the output is in the correct range.
> 
> [sample code]
> 

Making sure the output is in the correct range is what this patch is about, not the precision (in fact against precision in the latest version) or optimization. The sample code exhibits the very same bug I'm trying to squash here. Test with gain = 1.0f and someValue anything between 2147483584 and 2147483647. Or whatever values that should give output in that range for the gain you choose.

Patch-set: 2
Reviewer: Gerrit User 1000223 <1000223@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>