Update patch set 1
Patch Set 1:

Please note the context is in https://dev.haiku-os.org/ticket/8316#comment:19. I'm not sure whether it's possible to add a Gerrit flag to attach a relevant bug; if there is, it would be useful to do so.

I'm not necessarily suggesting this is the best way to solve the problem: the above bug comment has some alternatives which are worth considering before committing this change.

Any and all feedback welcome.

Patch-set: 1
Reviewer: Gerrit User 1000268 <1000268@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>