Update patch set 1
Patch Set 1: Code-Review+2

> Please note the context is in https://dev.haiku-os.org/ticket/8316#comment:19. I'm not sure whether it's possible to add a Gerrit flag to attach a relevant bug; if there is, it would be useful to do so.

You can mention it in the commit message, something like this:

Fixes #8316

Gerrit will automatically link to the bugtracker then.

> I'm not necessarily suggesting this is the best way to solve the problem: the above bug comment has some alternatives which are worth considering before committing this change.

Looks good to me. That's what this flag is meant to do: allow using addresses that are not reachable yet because the network interface isn't configured in that address family. The default is to not do that and avoid accidentally trying to reach an IPv6 address when locally IPv6 is disabled.

A further improvement would be to add the missing error checks in BNetworkAddress, so that it doesn't end up with random data but instead something more clearly invalid (set everything to 0 or something like that?).

Patch-set: 1
Reviewer: Gerrit User 1000001 <1000001@40b9299a-d8a8-485d-9b01-e6d3f45eefb5>
Label: Code-Review=+2